# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /post:
    # our controller name
    x-swagger-router-controller: post
    get:
      operationId: getPostList
      description: get the posts list
      # define the type of response for Success "200" and Error
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetPostListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: createPost
      description: add a new post to the list
      # post info to be stored
      parameters:
        - name: body
          description: Post properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Post"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  GetPostListResponse:
    required:
      - posts
    properties:
      # The array of movies
      posts:
        type: array
        items: 
          type: object
          properties:
            _id:
              type: string
            username:
              type: string
            text:
              type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  Post:
    type: object
    properties:
      username:
        type: string
        description: post author name
      text:
        type: string
        description: post text
    required:
      - username
      - text
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description